feature/dev: add Public module
